Add OFI (libfabric) RDMA backend
Add OFIContext and OFIRemoteMemoryAgent implementing the
RemoteMemoryAgent interface over libfabric. Supports RDM endpoints
with automatic provider selection (prefers CXI/EFA/ofi_rxm, skips
ofi_rxd and shm), offset-based and virtual addressing modes, and
completion-tracked writes/reads/atomics.

RMAFactory gains OFI_RDMA and AUTO_RDMA now prefers OFI > IBV > MPI.
RemoteBufferInfo::rkey widened to uint64_t for OFI key sizes.

+++ RemoteMemoryAgent.hpp
@@ -30,7 +30,7 @@ namespace rma_detail {
 struct RemoteBufferInfo
 {
     uint64_t addr = 0;
-    uint32_t rkey = 0;
+    uint64_t rkey = 0;
     size_t count = 0;
 };
